requirements.txt
smartsheet-python-sdk==3.0.4
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
setuptools==69.5.1
reportlab==4.0.4
svglib==1.5.1
qrcode==7.4.2
Pillow==10.4.0
dropbox==11.36.2